move verify message and map to curve to bn254, nits
