+++ crates/chainio/utils/Cargo.toml
@@ -9,8 +9,3 @@ repository.workspace = true
 license-file.workspace = true
 
 [dependencies]
-ark-bn254 = "0.4.0"
-ark-ff.workspace = true
-eigen-crypto-bls.workspace = true
-eigen-crypto-bn254.workspace = true
-eigen-utils.workspace = true
