+++ crates/chainio/utils/src/lib.rs
@@ -2,4 +2,3 @@
     html_logo_url = "https://github.com/Layr-Labs/eigensdk-rs/assets/91280922/bd13caec-3c00-4afc-839a-b83d2890beb5",
     issue_tracker_base_url = "https://github.com/Layr-Labs/eigensdk-rs/issues/"
 )]
-// #![cfg_attr(not(test), warn(unused_crate_dependencies))]
