+++ crates/crypto/bls/Cargo.toml
@@ -9,17 +9,14 @@ repository.workspace = true
 license-file.workspace = true
 
 [dependencies]
-ark-bn254 = "0.4.0"
+ark-bn254.workspace = true
 ark-ff.workspace = true
 eigen-crypto-bn254.workspace = true
 thiserror.workspace = true
-ark-ec = "0.4.2"
+ark-ec.workspace = true
 alloy-primitives.workspace = true
 eigen-utils.workspace = true
-num-bigint = "0.4.4"
 ark-std = { version = "0.4.0", default-features = false }
-sha2 = { version = "0.10.6", default-features = false }
-rust-bls-bn254.workspace = true
 [dev-dependencies]
 rand = "0.8.4"
 tokio = { workspace = true, features = ["full"] }
