+++ crates/crypto/bn254/Cargo.toml
@@ -9,10 +9,12 @@ repository.workspace = true
 license-file.workspace = true
 
 [dependencies]
+ark-bn254.workspace = true
+ark-ec.workspace = true
 ark-ff.workspace = true
-ark-bn254 = "0.4.0"
-thiserror.workspace = true
-alloy-primitives.workspace = true
+num-bigint = "0.4.4"
+rust-bls-bn254.workspace = true
+
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["full"] }
