+++ crates/crypto/bn254/src/error.rs
@@ -1,8 +1,1 @@
-use thiserror::Error;
 
-#[derive(Debug, Error)]
-pub enum Bn254Err {
-    /// Build Fq instance from string
-    #[error("failed to build a Fq instance from string ")]
-    Fq,
-}
