+++ examples/avsregistry-write/examples/register_operator_in_quorum_with_avs_registry_coordinator.rs
@@ -13,7 +13,14 @@ use eigen_testing_utils::m2_holesky_constants::{
 };
 use eigen_types::operator::Operator;
 use eyre::Result;
+use lazy_static::lazy_static;
 use std::str::FromStr;
+use std::time::{SystemTime, UNIX_EPOCH};
+
+lazy_static! {
+    /// 1 day
+    static ref SIGNATURE_EXPIRY: U256 = U256::from(86400);
+}
 #[tokio::main]
 async fn main() -> Result<()> {
     let holesky_provider = "https://ethereum-holesky.blockpi.network/v1/rpc/public";
@@ -28,6 +35,8 @@ async fn main() -> Result<()> {
     )
     .await
     .expect("avs writer build fail ");
+
+    // Create a new key pair instance using the secret key
     let bls_key_pair = BlsKeyPair::new(
         "12248929636257230549931416853095037629726205319386239410403476017439825112537".to_string(),
     )
@@ -38,9 +47,23 @@ async fn main() -> Result<()> {
         0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02,
         0x02, 0x02,
     ]);
-    let sig_expiry: U256 = U256::from(1723123419);
+
+    // Get the current SystemTime
+    let now = SystemTime::now();
+    let mut sig_expiry: U256 = U256::from(0);
+    // Convert SystemTime to a Duration since the UNIX epoch
+    if let Ok(duration_since_epoch) = now.duration_since(UNIX_EPOCH) {
+        // Convert the duration to seconds
+        let seconds = duration_since_epoch.as_secs(); // Returns a u64
+
+        // Convert seconds to U256
+        sig_expiry = U256::from(seconds) + *SIGNATURE_EXPIRY;
+    } else {
+        println!("System time seems to be before the UNIX epoch.");
+    }
     let quorum_nums = Bytes::from([0x01]);
 
+    // A new ElChainReader instance
     let el_chain_reader = ELChainReader::new(
         get_test_logger().clone(),
         SLASHER_ADDRESS,
@@ -48,6 +71,7 @@ async fn main() -> Result<()> {
         AVS_DIRECTORY_ADDRESS,
         "https://ethereum-holesky.blockpi.network/v1/rpc/public".to_string(),
     );
+    // A new ElChainWriter instance
     let el_writer = ELChainWriter::new(
         DELEGATION_MANAGER_ADDRESS,
         STRATEGY_MANAGER_ADDRESS,
@@ -55,10 +79,12 @@ async fn main() -> Result<()> {
         "https://ethereum-holesky.blockpi.network/v1/rpc/public".to_string(),
         "bead471191bea97fc3aeac36c9d74c895e8a6242602e144e43152f96219e96e8".to_string(),
     );
+
     let wallet = PrivateKeySigner::from_str(
         "bead471191bea97fc3aeac36c9d74c895e8a6242602e144e43152f96219e96e8",
     )
     .expect("no key ");
+
     let operator_details = Operator::new(
         wallet.address(),
         wallet.address(),
@@ -66,15 +92,17 @@ async fn main() -> Result<()> {
         3,
         Some("eigensdk-rs".to_string()),
     );
+    // Register the address as operator in delegation manager
     let _s = el_writer.register_as_operator(operator_details).await;
 
+    // Register the operator in registry coordinator
     avs_registry_writer
         .register_operator_in_quorum_with_avs_registry_coordinator(
             bls_key_pair,
             digest_hash,
             sig_expiry,
             quorum_nums,
-            "65.109.158.181:33078;31078".to_string(),
+            "65.109.158.181:33078;31078".to_string(), // socket
         )
         .await
         .unwrap();
